refactor: 도메인 로직과 Rule에 관련된 로직 분리
