+++ src/test/kotlin/omok/model/model/TurnTest.kt
@@ -1,5 +1,11 @@
-package omok.model
+package omok.model.model
 
+import omok.model.BlackTurn
+import omok.model.Board
+import omok.model.Point
+import omok.model.Stone
+import omok.model.StoneType
+import omok.model.WhiteTurn
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
