+++ src/test/kotlin/omok/model/rule/RenjuRuleTest.kt
@@ -1,5 +1,10 @@
-package omok.model
+package omok.model.rule
 
+import omok.model.BlackTurn
+import omok.model.Board
+import omok.model.Point
+import omok.model.Stone
+import omok.model.StoneType
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
