+++ src/test/kotlin/omok/model/rule/RuleTest.kt
@@ -1,5 +1,11 @@
-package omok.model
+package omok.model.rule
 
+import omok.model.Board
+import omok.model.Point
+import omok.model.Stone
+import omok.model.StoneType
+import omok.rule.BlackRule
+import omok.rule.WhiteRule
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
